Add modal to create a new worker and invoice
refs: fga-eps-mds/2018.2-PDF2CASH#151
