+++ pdf2cash_front/pages/invoice/index.js
@@ -11,6 +11,9 @@ import DeleteIcon from '@material-ui/icons/Delete';
 import VisibilityIcon from '@material-ui/icons/Visibility';
 import { withStyles } from '@material-ui/core/styles';
 import CustomatizedTable from '../../comps/table';
+import AddIcon from '@material-ui/icons/Add';
+import InvoiceCreate from '../../comps/createInvoice'
+import Authenticate from '../auth';
 
 const styles = theme => ({
   cell: {
@@ -20,7 +23,7 @@ const styles = theme => ({
     textAlign: 'center',
     marginTop: 100,
   },
-  paper: {
+  paperDelete: {
     position: 'absolute',
     width: theme.spacing.unit * 50,
     backgroundColor: theme.palette.background.paper,
@@ -28,7 +31,15 @@ const styles = theme => ({
     padding: theme.spacing.unit * 4,
     textAlign: 'center',
   },
-  buttom: {
+  paperCreate: {
+    position: 'absolute',
+    width: theme.spacing.unit * 70,
+    backgroundColor: theme.palette.background.paper,
+    boxShadow: theme.shadows[5],
+    padding: theme.spacing.unit * 4,
+    textAlign: 'center'
+  },
+  button: {
     marginTop: 30,
   },
 });
@@ -51,7 +62,8 @@ class InvoiceIndex extends Component {
       invoices: [],
       sellers: [],
       join: [],
-      open: false,
+      openDelete: false,
+      openCreate: false,
       id: 0,
     };
 
@@ -59,17 +71,27 @@ class InvoiceIndex extends Component {
     this.dateFormatter = this.dateFormatter.bind(this);
     this.delete = this.delete.bind(this);
     this.getInvoices = this.getInvoices.bind(this);
-    this.openModal = this.openModal.bind(this);
-    this.closeModal = this.closeModal.bind(this);
+    this.openModalDelete = this.openModalDelete.bind(this);
+    this.closeModalDelete = this.closeModalDelete.bind(this);
+    this.openModalCreate = this.openModalCreate.bind(this);
+    this.closeModalCreate = this.closeModalCreate.bind(this);
   }
 
   async componentDidMount() {
     const urlInvoice = 'http://localhost:8008/api/invoice/invoice/';
-    const resInvoice = await fetch(urlInvoice);
+    const resInvoice = await fetch(urlInvoice,
+      {
+        method: 'GET',
+        headers: {
+          'Content-Type': 'application/json',
+          'Authorization': 'JWT ' + Authenticate.getToken()
+        },
+        credentials: 'omit',
+      });
     const dataInvoice = await resInvoice.json();
     this.setState({
       invoices: dataInvoice,
-      open: false,
+      openDelete: false,
     });
 
     const urlSeller = 'http://localhost:8008/api/invoice/seller/';
@@ -129,16 +151,28 @@ class InvoiceIndex extends Component {
     this.setState({ join });
   }
 
-  openModal(id) {
+  openModalDelete(id) {
     this.setState({
-      open: true,
+      openDelete: true,
       id,
     });
   }
 
-  async closeModal() {
+  async closeModalDelete() {
     this.setState({
-      open: false,
+      openDelete: false,
+    });
+  }
+
+  openModalCreate() {
+    this.setState({
+      openCreate: true,
+    });
+  }
+
+  async closeModalCreate() {
+    this.setState({
+      openCreate: false
     });
   }
 
@@ -146,7 +180,8 @@ class InvoiceIndex extends Component {
     const { classes } = this.props;
     const {
       join,
-      open,
+      openDelete,
+      openCreate,
       id,
       invoices,
     } = this.state;
@@ -156,32 +191,53 @@ class InvoiceIndex extends Component {
             <Typography variant="display2">
               Listar Notas Fiscais
             </Typography>
+            <Button
+              variant="fab"
+              color="secondary"
+              aria-label="Add"
+              className={classes.buttonCreate}
+              onClick={() => this.openModalCreate()}
+            >
+              <AddIcon />
+            </Button>
             <Modal
               aria-labelledby="simple-modal-title"
               aria-describedby="simple-modal-description"
-              open={ open }
-              onClose={ this.closeModal }
+              open={ openDelete }
+              onClose={ this.closeModalDelete }
             >
-                <Grid style={ getModalStyle() } className={ classes.paper }>
+                <Grid style={ getModalStyle() } className={ classes.paperDelete }>
                     <Typography variant="h6" className={ classes.cell }>
-                      Deseja realmete deletar essa nota fiscal ?
+                      Deseja realmente deletar essa nota fiscal ?
                     </Typography>
                     <Button
+                      id='SIM'
                       variant="contained"
-                      className={ classes.buttom }
+                      className={ classes.button }
                       color="primary"
                       onClick={ () => this.delete(id) }
                     >
                       SIM
                     </Button>
                     <Button
+                      id='NAO'
                       variant="contained"
-                      className={ classes.buttom }
+                      className={ classes.button }
                       color="secondary"
-                      onClick={ this.closeModal }
+                      onClick={ this.closeModalDelete }
                     >
                       NÃO
                     </Button>
+                </Grid>
+            </Modal>
+            <Modal
+              aria-labelledby="simple-modal-title"
+              aria-describedby="simple-modal-description"
+              open={ openCreate }
+              onClose={ this.closeModalCreate }
+            >
+                <Grid style={ getModalStyle() } className={ classes.paperCreate }>
+                  < InvoiceCreate />
                 </Grid>
             </Modal>
             {
@@ -221,7 +277,7 @@ class InvoiceIndex extends Component {
                             </Button>
                         </TableCell>
                         <TableCell className={ classes.cell }>
-                            <Button onClick={ () => this.openModal(invoice.id) }>
+                            <Button onClick={ () => this.openModalDelete(invoice.id) }>
                                 <DeleteIcon />
                             </Button>
                         </TableCell>
